Rename 'icon' to 'generic-icon' in the mime-info file
The Freedesktop theme spec says that icons are first searched in
the current theme, then in its parents and finally in the default
"hicolor" theme.

"hicolor" is where Fuse installs its icons, however if generic-icon is
not specified in fuse.xml then it defaults to application-x-generic.

This means that if the current theme provides an icon named
application-x-generic it will be used instead of the ZX Spectrum one.

Fixes bug #494